refactor: move test to match previous refactoring
